distributed_video/requirements.txt
Flask==2.2.2
python-decouple
opencv-python
celery
requests
pre-commit
pyzmq
